requirements.txt
requests
python-dotenv
ipwhois
ipinfo